Refactor Invoke-TelemetryDisable function to check for existing AtlasModules before downloading and installing

+++ functions/Invoke-DisableDefender.ps1
@@ -1,8 +1,12 @@
 function Invoke-DisableDefender {
 
-	Invoke-WebRequest -Uri "https://github.com/ManuelBiscotti/test/raw/refs/heads/main/tools/AtlasModules.zip" -OutFile "$env:TEMP\AtlasModules.zip"
-	Expand-Archive -Path "$env:TEMP\AtlasModules.zip" -DestinationPath "$env:TEMP" -Force -ErrorAction SilentlyContinue
-	Move-Item -Path (Join-Path $env:TEMP 'AtlasModules') -Destination 'C:\Windows' -Force -ErrorAction SilentlyContinue
-    Start-Process -FilePath 'C:\Windows\AtlasModules\DisableDefender.cmd' -NoNewWindow -Wait
-
+    if (Test-Path "C:\Windows\AtlasModules") {
+        Start-Process -FilePath 'C:\Windows\AtlasModules\DisableDefender.cmd' -NoNewWindow -Wait
+    }else {
+ 	    Invoke-WebRequest -Uri "https://github.com/ManuelBiscotti/test/raw/refs/heads/main/tools/AtlasModules.zip" -OutFile "$env:TEMP\AtlasModules.zip"
+	    Expand-Archive -Path "$env:TEMP\AtlasModules.zip" -DestinationPath "$env:TEMP" -Force -ErrorAction SilentlyContinue
+	    Move-Item -Path (Join-Path $env:TEMP 'AtlasModules') -Destination 'C:\Windows' -Force -ErrorAction SilentlyContinue
+        Start-Process -FilePath 'C:\Windows\AtlasModules\DisableDefender.cmd' -NoNewWindow -Wait       
+    }
+    
 }
